+++ generators/cucumber/templates/env.rb
@@ -24,6 +24,9 @@ After do |scenario|
     DataStorage.store('sauce', :status, false) if scenario.failed?
     session_end = duration(Time.now.utc - DataStorage.extract('sauce', :start_time))
     log.info "SAUCE VIDEO #@session_start - #{session_end} URL: #{sauce_resource_path('video.flv')}"
+  elsif ie_browser?
+    log.info 'IE reset session'
+    page.execute_script("void(document.execCommand('ClearAuthenticationCache', false));")
   end
   DataStorage.clear_ns("user")
 end