+++ generators/rspec/templates/spec_helper.rb
@@ -35,13 +35,16 @@ RSpec.configure do |config|
     if sauce_driver?
       session_end = duration(Time.now.utc - DataStorage.extract('sauce', :start_time))
       log.info "SAUCE VIDEO #@session_start - #{session_end} URL: #{sauce_resource_path('video.flv')}"
+    elsif ie_browser?
+      log.info 'IE reset session'
+      page.execute_script("void(document.execCommand('ClearAuthenticationCache', false));")
     end  
   end
 
   config.after(:suite) do
     if sauce_driver?
       report_failures_count = config.reporter.instance_variable_get(:@failure_count)
-      DataStorage.store('sauce', :status, report_failures_count.zero?) 
+      DataStorage.store('sauce', :status, report_failures_count.zero?)
     end
   end